Background styles, folder expiration date
The back ground is now dynamic based on  the url and remained fixed as
the foreground scrolled.
The folder cleanup function get rid of folders after 31 days however
the user was still able to log in. Thus this is now fixed: the user can
no longer log in after a folder has been erased.